+++ src/main/java/net/frozenblock/lib/levelgen/feature/api/FrozenLibFeatures.java
@@ -24,15 +24,17 @@ import net.frozenblock.lib.levelgen.feature.api.feature.CircularWaterloggedVeget
 import net.frozenblock.lib.levelgen.feature.api.feature.CircularWaterloggedVegetationPatchLessBordersFeature;
 import net.frozenblock.lib.levelgen.feature.api.feature.ColumnFeature;
 import net.frozenblock.lib.levelgen.feature.api.feature.ColumnWithDiskFeature;
+import net.frozenblock.lib.levelgen.feature.api.feature.ConfigEntrySelectorFeature;
 import net.frozenblock.lib.levelgen.feature.api.feature.CurvingSpikeFeature;
 import net.frozenblock.lib.levelgen.feature.api.feature.LargeSpireFeature;
 import net.frozenblock.lib.levelgen.feature.api.feature.UnderwaterVegetationPatchFeature;
 import net.frozenblock.lib.levelgen.feature.api.feature.UnderwaterVegetationPatchWithEdgeDecorationFeature;
 import net.frozenblock.lib.levelgen.feature.api.feature.VegetationPatchWithEdgeDecorationFeature;
-import net.frozenblock.lib.levelgen.feature.api.feature.config.ColumnFeatureConfiguration;
-import net.frozenblock.lib.levelgen.feature.api.feature.config.ColumnWithDiskFeatureConfiguration;
-import net.frozenblock.lib.levelgen.feature.api.feature.config.CurvingSpikeConfig;
-import net.frozenblock.lib.levelgen.feature.api.feature.config.LargeSpireConfig;
+import net.frozenblock.lib.levelgen.feature.api.feature.configurations.ColumnFeatureConfiguration;
+import net.frozenblock.lib.levelgen.feature.api.feature.configurations.ColumnWithDiskFeatureConfiguration;
+import net.frozenblock.lib.levelgen.feature.api.feature.configurations.ConfigEntrySelectorFeatureConfiguration;
+import net.frozenblock.lib.levelgen.feature.api.feature.configurations.CurvingSpikeConfiguration;
+import net.frozenblock.lib.levelgen.feature.api.feature.configurations.LargeSpireConfiguration;
 import net.frozenblock.lib.levelgen.feature.api.feature.disk.BallFeature;
 import net.frozenblock.lib.levelgen.feature.api.feature.disk.config.BallFeatureConfiguration;
 import net.frozenblock.lib.levelgen.feature.api.feature.noise_path.NoisePathFeature;
@@ -42,6 +44,7 @@ import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.world.level.levelgen.feature.configurations.VegetationPatchConfiguration;
 
 public class FrozenLibFeatures {
+	public static final ConfigEntrySelectorFeature CONFIG_ENTRY_SELECTOR = new ConfigEntrySelectorFeature(ConfigEntrySelectorFeatureConfiguration.CODEC);
 	public static final NoisePathFeature NOISE_PATH = new NoisePathFeature(NoisePathFeatureConfiguration.CODEC);
 	public static final BallFeature BALL = new BallFeature(BallFeatureConfiguration.CODEC);
 	public static final ColumnFeature COLUMN = new ColumnFeature(ColumnFeatureConfiguration.CODEC);
@@ -52,10 +55,11 @@ public class FrozenLibFeatures {
 	public static final CircularWaterloggedVegetationPatchLessBordersFeature CIRCULAR_WATERLOGGED_VEGETATION_PATCH_LESS_BORDERS = new CircularWaterloggedVegetationPatchLessBordersFeature(VegetationPatchConfiguration.CODEC);
 	public static final CircularLavaVegetationPatchFeature CIRCULAR_LAVA_VEGETATION_PATCH = new CircularLavaVegetationPatchFeature(VegetationPatchConfiguration.CODEC);
 	public static final CircularLavaVegetationPatchLessBordersFeature CIRCULAR_LAVA_VEGETATION_PATCH_LESS_BORDERS = new CircularLavaVegetationPatchLessBordersFeature(VegetationPatchConfiguration.CODEC);
-	public static final LargeSpireFeature LARGE_SPIRE = new LargeSpireFeature(LargeSpireConfig.CODEC);
-	public static final CurvingSpikeFeature CURVING_SPIKE = new CurvingSpikeFeature(CurvingSpikeConfig.CODEC);
+	public static final LargeSpireFeature LARGE_SPIRE = new LargeSpireFeature(LargeSpireConfiguration.CODEC);
+	public static final CurvingSpikeFeature CURVING_SPIKE = new CurvingSpikeFeature(CurvingSpikeConfiguration.CODEC);
 
 	public static void init() {
+		Registry.register(BuiltInRegistries.FEATURE, FrozenLibConstants.id("config_entry_selector"), CONFIG_ENTRY_SELECTOR);
 		Registry.register(BuiltInRegistries.FEATURE, FrozenLibConstants.id("noise_path"), NOISE_PATH);
 		Registry.register(BuiltInRegistries.FEATURE, FrozenLibConstants.id("ball"), BALL);
 		Registry.register(BuiltInRegistries.FEATURE, FrozenLibConstants.id("column"), COLUMN);