+++ src/main/java/net/frozenblock/lib/levelgen/feature/api/feature/ColumnFeature.java
@@ -18,7 +18,7 @@
 package net.frozenblock.lib.levelgen.feature.api.feature;
 
 import com.mojang.serialization.Codec;
-import net.frozenblock.lib.levelgen.feature.api.feature.config.ColumnFeatureConfiguration;
+import net.frozenblock.lib.levelgen.feature.api.feature.configurations.ColumnFeatureConfiguration;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
 import net.minecraft.util.RandomSource;