+++ CHANGELOG.md
@@ -40,6 +40,11 @@ Put the changelog BELOW the dashes. ANYTHING ABOVE IS IGNORED.
 - Added the `ConfigEntryCondition`.
   - A wrapper of `ConfigEntryPredicate` in `LootItemCondition` form.
   - Use the `asLootCondition` method in `ConfigEntryPredicate`. Failing to do so will cause unnecessary memory usage.
+- Added the `ConfigEntrySelector` Configured Feature.
+  - Is configured with `ConfigEntrySelectorFeatureConfiguration` with the following fields:
+    - `config_entry_predicate`: The `ConfigEntryPredicate` to use.
+    - `feature_if_true`: The `PlacedFeature` to be placed if `config_entry_predicate` is true.
+    - `feature_if_false`: The `PlacedFeature` to be placed if `config_entry_predicate` is false.
 - Refactored and renamed many worldgen-related classes to be more consistent with Vanilla.
 - `ColumnWithDiskFeatureConfiguration` now uses two `BlockStateProvider`s instead of a `BlockState` and Block Tag.
 - Renamed the `block_state_provider` field to `block_state` and the `stop_when_encountering_unreplaceable_block` field to `stop_at_unreplaceable_block` in `ColumnFeatureConfiguration`.